pyproject.toml
[project]
name = "eiffel-steering"
version = "0.1.0"
description = "Steering a large language model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.8.0",
    "sae-lens>=6.12.3",
    "transformers>=4.56.2",
    "nnsight>=0.5.5",
    "datasets>=4.1.1",
    "hf-transfer>=0.1.9",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
    "botorch>=0.15.1",
    "httpx>=0.27",
    "pydantic>=2.9",
    "typer>=0.12",
    "rich>=13.9",
    "questionary>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
persona-rerank = ["sentence-transformers>=3.0"]
test = ["pytest>=8.0"]

[project.scripts]
persona = "persona.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["persona*"]